Image and audio categories (#13)
Add options for Text, Image, or Audio when creating categories. If Image or Audio is selected, then we skip the query to ChatGPT and instead produce five Q-A pairs with blank question/answer fields. These are intended to be populated with a file path to an image or audio file and the answer to that clue, manually entered by the user.

The user can also manually alter individual Q-A types and this will be reflected in the output text file.

---------

Co-authored-by: Joshua Rea <josh.rea@gmail.com>